Create api.js
Adding the routes/api.js file to the project's repo. Written the logic code for: 1. adding a new stock object to the DB; 2. checking to see if it's already in the DB; 3. Updating the entry in the DB if found; 4. Adding likes only once per ip. Always the result get's the price value updated in real time from the alphavantage API so no old prices from the DB. Only need to add the possibility to compare 2 stocks in real time and to write the functional_tests to complete the project's user stories.