Create stockHandler.js
Very important to note is that the FCC's proposed google.finance API has been discontinued and as an alternative I've found alphavantage.co stock information API that's responsive and has a functional library on npm. I've written and separated the stock price checking and stock compare logic into 2 handlers: stockHandler.js and compareHandler.js. As in past projects, I found that using modularity in the code helps in reading and testing the code easier. The stockHandler.js handles the first form, where only one singular stock is requested and the compareHandler.js is handling the second form, where the 2 concurrent stocks are requested and compared.