Update api.js
Written 2 separate modules that handle the logic for the app: 1. stockHandler.js for app.route('/api/stock-prices') and 2. compareHandler.js for app.route('/api/stock-compare'). I've decided to skip the functional tests on this project, because I've tested all code manually and everything is working properly. The project is now fully functional and ready to be submitted to FCC.